projectplanning-app/cloud-api/pyproject.toml
[project]
name = "projectplanning-api"
version = "0.1.0"
description = ""
authors = [
    {name = "juampi",email = "jpmiya82@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
